Send email on successful sign up
Once the data has been written out and there are no issues, send the
participant a confirmation email. Also, if this is a multi-person
experiment and the slot is now full, send emails to the other people in
the slot to let them know that the slot has been filled.