Deposit cookie to prevent excluded participants
If a user is told they’re ineligible, a cookie is deposited on their
system identifying the experiment ID. If they then try again to sign up
for the experiment again, even with a different email address, they’ll
continue to be told that they’re ineligible. The cookie will expire
after one week.